+++ R/kg_join.R
@@ -20,21 +20,21 @@
 #' @return A `tbl_kgx()` graph
 #' @export
 #' @examples
-#' ## Using local MONDO KGX file (packaged with monarchr)
-#' monarch <- file_engine(system.file("extdata", "mondo_kgx_tsv.tar.gz",
-#'                               package = "monarchr"))
+#' ## Using example KGX file packaged with monarchr
+#' filename <- system.file("extdata", "eds_marfan_kg.tar.gz", package = "monarchr")
+#' engine <- file_engine(filename)
 #'
-#' eds_and_phenos <- monarch |>
+#' eds_and_phenos <- engine |>
 #'                   fetch_nodes(query_ids = "MONDO:0007525") |>
 #'                   expand(predicates = "biolink:has_phenotype",
 #'                          categories = "biolink:PhenotypicFeature")
 #'
-#' marfan_and_phenos <- monarch |>
+#' marfan_and_phenos <- engine |>
 #'                      fetch_nodes(query_ids = "MONDO:0007947") |>
 #'                      expand(predicates = "biolink:has_phenotype",
 #'                             categories = "biolink:PhenotypicFeature")
 #'
-#' combined <- graph_join(eds_and_phenos, marfan_and_phenos)
+#' combined <- kg_join(eds_and_phenos, marfan_and_phenos)
 #' print(combined)
 kg_join <- function(graph1, graph2, ...) {
   UseMethod("kg_join")